src/main/java/az/micro/msscbreweryclient/web/client/BreweryClient.java
package az.micro.msscbreweryclient.web.client;

import az.micro.msscbreweryclient.web.model.BeerDto;
import az.micro.msscbreweryclient.web.model.CustomerDto;
import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.boot.web.client.RestTemplateBuilder;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestTemplate;

import java.net.URI;
import java.util.UUID;

@Component
@ConfigurationProperties(prefix = "sfg.brewery" , ignoreUnknownFields = false)
public class BreweryClient {

    public final String BEER_PATH_V1 = "/api/v1/beer/";
    public final String CUSTOMER_PATH_V1 = "/api/v1/customer/";
    private String apiHost;
    private final RestTemplate restTemplate;

    public BreweryClient (RestTemplateBuilder restTemplateBuilder) {
        this.restTemplate = restTemplateBuilder.build();
    }

    public CustomerDto getCustomerById (UUID uuid) {
        return restTemplate.getForObject(apiHost + CUSTOMER_PATH_V1 + uuid.toString() , CustomerDto.class);
    }

    public CustomerDto saveNewCustomer () {
        return restTemplate.getForObject(apiHost + CUSTOMER_PATH_V1 , CustomerDto.class);
    }

    public void deleteCustomer (UUID uuid) {
        restTemplate.delete(apiHost + CUSTOMER_PATH_V1 + uuid.toString());
    }

    public void updateCustomer (UUID uuid , CustomerDto customerDto) {
        restTemplate.put(apiHost + CUSTOMER_PATH_V1 + uuid.toString() , customerDto);
    }

    public BeerDto getBeerById (UUID uuid) {
        return restTemplate.getForObject(apiHost + BEER_PATH_V1 + uuid.toString(), BeerDto.class);
    }

    public URI saveNewBeerDto (BeerDto beerDto) {
        return restTemplate.postForLocation(apiHost + BEER_PATH_V1 , beerDto);
    }

    public void updateBeer (UUID uuid , BeerDto beerDto) {
       restTemplate.put(apiHost + BEER_PATH_V1 + "/" + uuid.toString(), beerDto);
    }

    public void deleteBeer (UUID uuid) {
        restTemplate.delete(apiHost + BEER_PATH_V1 + "/" + uuid.toString());
    }

    public void setApiHost(String apiHost) {
        this.apiHost = apiHost;
    }
}
