src/test/java/az/micro/msscbreweryclient/web/client/BreweryClientTest.java
package az.micro.msscbreweryclient.web.client;

import az.micro.msscbreweryclient.web.model.BeerDto;
import az.micro.msscbreweryclient.web.model.CustomerDto;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import java.net.URI;
import java.util.UUID;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
class BreweryClientTest {

    @Autowired
    BreweryClient client;


    @Test
    void getCustomerById () {
        CustomerDto dto = client.getCustomerById(UUID.randomUUID());
        assertNotNull(dto);
    }
    @Test
    void testUpdateCustomer () {
        CustomerDto dto = CustomerDto.builder().name("Ulvi Rustamzade").build();
        client.updateCustomer(UUID.randomUUID(),dto);

    }
    @Test
    void deleteCustomer () {
        client.deleteCustomer(UUID.randomUUID());
    }

    @Test
    void getBeerById() {
        BeerDto dto = client.getBeerById(UUID.randomUUID());

        assertNotNull(dto);
    }

    @Test
    void testSavingBeer() {
        BeerDto beerDto = BeerDto.builder().beerName("New Beer").build();

        URI uri = client.saveNewBeerDto(beerDto);

        assertNotNull(uri);

        System.out.println(uri.toString());
    }

    @Test
    void testUpdateBeer() {
        BeerDto beerDto = BeerDto.builder().beerName("New Beer").build();

        client.updateBeer(UUID.randomUUID() , beerDto);
    }

    @Test
    void deleteBeer() {
        client.deleteBeer(UUID.randomUUID());
    }
}